requirements.txt
customtkinter==5.2.2
firebase-admin==6.2.0
Pillow==10.1.0
pyinstaller==6.1.0
